Closes #30
When a player bullet collides with an enemy bullet, both are removed.
This condition is checked with a double loop, the outer one over the
player's bullets, and the inner one over the enemies bullets.

Before the change, after a collision had been detected for a given enemy
and player bullets, both would be removed, but the inner loop would
continue to check if the player bullet (the one that was removed), also
collided with another enemy bullet. If another match was found, as
descrived in issue #30, the loop would try to remove the player bullet,
again, which is a no-no.

The break fixes that, since it stops the inner loop, which goes over the
enemy bullets